优化数据，单位mA mV 保留小数点后五位

+++ Forms/keithley2306/PowerData.cs
@@ -4,22 +4,22 @@ namespace HW_Thermal_Tools.Forms.keithley2306
 {
     public class PowerData
     {
-        public float Current { get; set; }
-        public float CurrentMin { get; set; }
-        public float CurrentMax { get; set; }
-        public float CurrentAve { get; set; }
+        public double Current { get; set; }
+        public double CurrentMin { get; set; }
+        public double CurrentMax { get; set; }
+        public double CurrentAve { get; set; }
 
 
-        public float Voltage { get; set; }
-        public float VoltageMin { get; set; }
-        public float VoltageMax { get; set; }
-        public float VoltageAve { get; set; }
+        public double Voltage { get; set; }
+        public double VoltageMin { get; set; }
+        public double VoltageMax { get; set; }
+        public double VoltageAve { get; set; }
 
 
-        public float Power { get; set; }
-        public float PowerMin { get; set; }
-        public float PowerMax { get; set; }
-        public float PowerAve { get; set; }
+        public double Power { get; set; }
+        public double PowerMin { get; set; }
+        public double PowerMax { get; set; }
+        public double PowerAve { get; set; }
 
 
         public List<DataPoint> CurrentHistory { get; set; } = new List<DataPoint>(); // 电流历史记录列表
@@ -35,9 +35,9 @@ namespace HW_Thermal_Tools.Forms.keithley2306
     public class DataPoint
     {
         public DateTime Time { get; set; } // 时间字段，用于横轴数据
-        public float Value { get; set; } // 数值字段，用于纵轴数据
+        public double Value { get; set; } // 数值字段，用于纵轴数据
 
-        public DataPoint(DateTime time, float value) // 构造函数
+        public DataPoint(DateTime time, double value) // 构造函数
         {
             Time = time;
             Value = value;
@@ -47,11 +47,11 @@ namespace HW_Thermal_Tools.Forms.keithley2306
     public class ExcelDataPoint
     {
         public DateTime Time { get; set; } //时间
-        public float CurentValue { get; set; } //current value
-        public float VoltageValue { get; set; } //voltage value
-        public float PowerValue { get; set; } //power value
+        public double CurentValue { get; set; } //current value
+        public double VoltageValue { get; set; } //voltage value
+        public double PowerValue { get; set; } //power value
 
-        public ExcelDataPoint(DateTime time, float currentvalue, float voltageValue, float powerValue)
+        public ExcelDataPoint(DateTime time, double currentvalue, double voltageValue, double powerValue)
         {
             Time = time;
             VoltageValue = voltageValue;
